requirements.txt
bert-extractive-summarizer==0.10.1
blis==0.7.9
brotlipy==0.7.0
catalogue==2.0.8
certifi
cffi
charset-normalizer
click==8.1.3
colorama
confection==0.0.4
cryptography
cymem==2.0.7
filelock==3.10.0
Flask==2.2.3
Flask-Cors==3.0.10
germansentiment==1.1.0
huggingface-hub==0.13.2
idna
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.0
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pluggy
pycosat
pycparser
pydantic==1.10.6
pyOpenSSL
PySock
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
PyYAML==6.0
regex==2022.10.31
requests
ruamel.yaml.clib
scikit-learn==1.2.2
scipy==1.10.1
six
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
textblob==0.17.1
textblob-de==0.4.3
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.2
toolz
torch==2.0.0
tqdm
transformers==4.27.1
typer==0.7.0
typing_extensions==4.5.0
urllib3
wasabi==1.1.1
wcwidth==0.2.5
Werkzeug==2.2.3
win-inet-pton
wincertstore==0.2
zstandard==0.18.0
